fuse MFP to FlatMap
